Fix adding cache contexts only when it has an effect on visibility
Initial work on #154
eg. When a banner is set to show on specific pages, only add the cache context for url.path
when the current path is one of those pages.
When a banner is set to be hidden on specific pages, only add the cache context for url.path
when the current path is one of those pages, and skip on the rest.