Skip alert type postsave during config sync
Fix #251

During config sync operations, if creating a new alert banner type the
visibility field will try to be created, even though that is
part of config.

This checks if the isSyncing flag is set and skips over the
postSave actions as we assume they are part of the config
being imported.