Hide moderation controls when present on confirmation form page
Fix #236

Adds a preprocess implementation to the alert banner to hide the moderation
controls if they have been displayed when on the confirmation page
(status form).

Note: This only hides on the confirmation form page eg.
/admin/content/alert-banner/18/status

If the form is included in another way it will still be present.

This way, moderation controls can be placed on an alert banner manage display
if a site builder considers that desirable, without interfering with the
confirmation form to put an alert banner live.