Fix: Group banner publishing.
When a new banner is set to go live, we should end up in the "Put banner live"
tab of the new banner after the banner form is submitted.  At the moment we are
ending up in the banner's "View" tab.

Also, group banners should only appear in group pages and not all over the site.