Move the language selection logic to the block
This is to allow the entitiy to remain pure in terms of logic, and the get
translation logic in the block. If the alert has a translation use that in place
of the alert banner entitiy. This also allows correct checking of visibility

Also allows the title to be translated.