Opt into the Gin content form routes (edit theme)
Fix #229

Implements hook_gin_content_form_routes for both the alert banner creation
and edit form. This makes editing the alert banner like nodes.

Renames the publishing options group meta to auto pick up gin theming.
Moves the revision log and new revision into the meta group
Add a flag if this is the Gin theme, if it is add the extra class to meta
so it is themed like nodes.