Feature/201 remove jquery (#386)
* removes jQuery, adds modern (Drupal) JS

* adds once library as a dependency

* waits for button to be present before clicking it

* Remove DomcontentLoaded event

Since javascript is already in Drupal behaviours attach function, the dom
should already be present. Once behaviours should prevent multiple attachments

Moved the setAlertBannerHideCookie function outside the attach.

* CS: test fixes

* Restore find button by contents in alertbannerhidetest

* Add core/drupal and core/once dependencies

Fix tests and use of alert banner on themes that don't use the once in the
theme.

The new javascript no longer requires jQuery, but does require the use of
the core/drupal and core/once libraries in order to function. These are
provided by most themes, but not stark or stable9 which the tests run with.

Should also resolve any issues we have with forthcoming javascript library
changes. See:
  https://www.drupal.org/node/3473558
  Asset ordering is now more strictly determined by library dependencies

---------

Co-authored-by: Andy Broomfield <andy.broomfield@brighton-hove.gov.uk>